Projectile + HomingProjectile deflection + collision detection fix finished
Re-coded the projectile collision detection logic to ensure that:
If not Deflected, the projectile checks collision with the Hero.
If Deflected and no Kill, the projectile checks collision with nothing.
If Deflected and Kill, the projectile checks collision with its source (currently sourceRef overwrites HeroRef on deflect, but this could probably be commented out- I left it as-is just in case)