Finished the Simple Minimap (barring finetuning)
Finished the simpleminimap that draws everything as simple colored squares. This involved changes in almost every custom object to finish up its minimapObj to draw with the right color and alpha value (to actually be seen).
Changes to the jsons were done to make the minimap have different background colors and zooms than the main camera (and a small QoL tweak or two snuck into Level1).
Added a new array to the two levels to hold minimap versions of the terrain so it can be drawn. On the same note, added a drawMini to GameObject that does nothing so the full AllPlatform set can be drawn without dying when trying to draw a normal platform (which isn't a custom object, so it doesn't have a drawMini).