Kelvin moves to SpawnPoints when falling
When Kelvin falls off the map, he moves to a SpawnPoint (after taking some damage).
The movement of the camera is instant, so we need to interpolate it to the hero (which may require reworking wherever it is in the code that sets the cam to pan with the Hero).
The SpawnPoints are visible in the minimap as purple squares- this should be turned off for release!